+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # @moritzbrantner/charts
 
+## Unreleased
+
+### Patch Changes
+
+- Make the React 19 peer contract explicit by adding the React DOM peer dependency.
+
 ## 0.1.1
 
 ### Patch Changes